feat(GlobalExceptionHandler.java): add RestControllerAdvice to handle NoSuchProductException

+++ src/main/java/gift/service/ProductService.java
@@ -1,6 +1,7 @@
 package gift.service;
 
 import gift.dto.ProductDTO;
+import gift.exception.NoSuchProductException;
 import gift.repository.H2Repository;
 import java.util.Collection;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,7 +31,9 @@ public class ProductService {
     }
 
     public ProductDTO updateProduct(long id, ProductDTO productDTO) {
-        repository.getProduct(id);
+        if (repository.getProduct(id) == null) {
+            throw new NoSuchProductException();
+        }
         ProductDTO updatedProduct = new ProductDTO(id, productDTO.name(), productDTO.price(), productDTO.imageUrl());
         repository.updateProduct(updatedProduct);
         return updatedProduct;
@@ -38,6 +41,9 @@ public class ProductService {
 
     public ProductDTO deleteProduct(long id) {
         ProductDTO deletedProduct = repository.getProduct(id);
+        if (deletedProduct == null) {
+            throw new NoSuchProductException();
+        }
         repository.deleteProduct(id);
         return deletedProduct;
     }
